requirements.txt
aiofiles==22.1.0
aiosqlite==0.18.0
altair==4.2.2
altgraph==0.17.3
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.6.0
asttokens==2.0.5
attrs==21.4.0
Babel==2.12.1
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
blinker==1.5
build==0.10.0
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.5
commonmark==0.9.1
contourpy==1.0.7
cssselect==1.2.0
cycler==0.11.0
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.6
Django==4.1.7
django-ckeditor==6.5.1
django-js-asset==2.0.0
docopt==0.6.2
docutils==0.19
entrypoints==0.4
et-xmlfile==1.1.0
exceptiongroup==1.1.0
executing==0.8.3
fastjsonschema==2.15.3
filelock==3.9.0
Flask==2.2.3
flatbuffers==23.5.26
fonttools==4.38.0
fpdf==1.7.2
fqdn==1.5.1
future==0.18.2
gitdb==4.0.10
GitPython==3.1.31
idna==3.4
importlib-metadata==6.0.0
iniconfig==2.0.0
ipykernel==6.15.0
ipython==8.4.0
ipython-genutils==0.2.0
isoduration==20.11.0
itsdangerous==2.1.2
jaraco.classes==3.2.3
jedi==0.18.1
Jinja2==3.1.2
joblib==1.2.0
json5==0.9.11
jsonpointer==2.3
jsonschema==4.17.3
jupyter-events==0.6.3
jupyter-ydoc==0.2.3
jupyter_client==8.0.3
jupyter_core==5.3.0
jupyter_server==2.5.0
jupyter_server_fileid==0.8.0
jupyter_server_terminals==0.4.4
jupyter_server_ydoc==0.6.1
jupyterlab==3.6.1
jupyterlab-pygments==0.2.2
jupyterlab_server==2.20.0
keyring==23.13.1
kiwisolver==1.4.4
libclang==16.0.0
lxml==4.9.2
MarkupSafe==2.1.1
matplotlib==3.7.0
matplotlib-inline==0.1.3
mistune==0.8.4
more-itertools==9.0.0
mpmath==1.3.0
mysql-connector-python==8.0.33
nbclassic==0.5.3
nbclient==0.6.6
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.5
networkx==3.1
nltk==3.8.1
notebook==6.4.12
notebook_shim==0.2.2
numpy==1.23.5
oauthlib==3.2.2
opencv-python==4.7.0.72
openpyxl==3.1.1
packaging==21.3
pandas==1.5.3
pandocfilters==1.5.0
parsel==1.7.0
parso==0.8.3
pefile==2022.5.30
pickleshare==0.7.5
Pillow==9.4.0
pipenv==2022.12.19
pipreqs==0.4.11
pkginfo==1.9.6
platformdirs==2.6.2
plot==0.6.5
plotly==5.13.1
pluggy==1.0.0
prometheus-client==0.14.1
prompt-toolkit==3.0.30
protobuf==3.20.3
psutil==5.9.1
pure-eval==0.2.2
pyarrow==11.0.0
pyasn1==0.5.0
pycparser==2.21
pydeck==0.8.0
Pygments==2.12.0
pyinstaller==5.7.0
pyinstaller-hooks-contrib==2022.14
Pympler==1.0.1
pyparsing==3.0.9
pypiwin32==223
pyproject==1.3.1
pyproject_hooks==1.0.0
pyprojects==0.0.1
pyrsistent==0.18.1
PySimpleGUI==4.60.4
pytest==7.2.1
python-dateutil==2.8.2
python-json-logger==2.0.7
pytz==2022.4
pytz-deprecation-shim==0.1.0.post0
pywin32==306
pywin32-ctypes==0.2.0
pywinpty==2.0.10
PyYAML==6.0
pyzmq==25.0.1
readme-renderer==37.3
regex==2022.10.31
requests==2.28.2
requests-toolbelt==0.10.1
rfc3339-validator==0.1.4
rfc3986==2.0.0
rfc3986-validator==0.1.1
rich==13.0.1
scipy==1.10.1
selectorlib==0.16.0
semver==2.13.0
Send2Trash==1.8.0
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
stack-data==0.3.0
streamlit==1.18.1
sympy==1.12
tenacity==8.2.2
tensorboard-data-server==0.7.1
tensorflow-estimator==2.12.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
terminado==0.15.0
tinycss2==1.1.1
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
torch==2.0.1
tornado==6.2
tqdm==4.65.0
traitlets==5.9.0
twine==4.0.2
typing==3.7.4.3
typing_extensions==4.5.0
tzdata==2022.7
tzlocal==4.2
uri-template==1.2.0
urllib3==1.26.14
validators==0.20.0
virtualenv==20.17.1
virtualenv-clone==0.5.7
w3lib==2.1.1
watchdog==2.2.1
wcwidth==0.2.5
webcolors==1.12
webencodings==0.5.1
websocket-client==1.5.1
Werkzeug==2.2.3
wrapt==1.14.1
y-py==0.5.9
yarg==0.1.9
ypy-websocket==0.8.2
zipp==3.11.0
